[MAINT] Rename `engine` -> `core`
