+++ eidaws.stationlite/eidaws/stationlite/harvest/validate.py
@@ -6,7 +6,7 @@ import requests
 
 from urllib.parse import urlparse, urljoin
 
-from eidaws.stationlite.engine.utils import RestrictedStatus
+from eidaws.stationlite.core.utils import RestrictedStatus
 from eidaws.stationlite.harvest.request import binary_request
 from eidaws.stationlite.settings import STL_SERVICES
 from eidaws.utils.error import Error
@@ -70,12 +70,16 @@ validate_station_method_token = _validate_method_token
 validate_wfcatalog_method_token = _validate_method_token
 
 
-def validate_dataselect_method_token(url,
-        restricted_status=RestrictedStatus.OPEN):
+def validate_dataselect_method_token(
+    url, restricted_status=RestrictedStatus.OPEN
+):
     token = _get_method_token(url)
     if (
         token is None
-        or (restricted_status == RestrictedStatus.OPEN and token != FDSNWS_QUERY_METHOD_TOKEN)
+        or (
+            restricted_status == RestrictedStatus.OPEN
+            and token != FDSNWS_QUERY_METHOD_TOKEN
+        )
         or (
             restricted_status == RestrictedStatus.CLOSED
             and token != FDSNWS_QUERYAUTH_METHOD_TOKEN
@@ -86,8 +90,9 @@ def validate_dataselect_method_token(url,
         )
 
 
-def validate_availability_method_token(url,
-        restricted_status=RestrictedStatus.OPEN):
+def validate_availability_method_token(
+    url, restricted_status=RestrictedStatus.OPEN
+):
     token = _get_method_token(url)
     if (
         token is None
@@ -110,7 +115,9 @@ def validate_availability_method_token(url,
         )
 
 
-def validate_method_token(url, service, restricted_status=RestrictedStatus.OPEN):
+def validate_method_token(
+    url, service, restricted_status=RestrictedStatus.OPEN
+):
     """
     Validates the *service method token* AKA the *service resource*.
     """
